+++ hyracks-fullstack/hyracks/hyracks-dataflow-std/src/test/java/org/apache/hyracks/dataflow/std/structures/SerializableHashTableTest.java
@@ -46,7 +46,7 @@ public class SerializableHashTableTest {
     @Before
     public void setup() throws HyracksDataException {
         ctx = new FrameManager(256);
-        framePool = new DeallocatableFramePool(ctx, ctx.getInitialFrameSize() * 2048);
+        framePool = new DeallocatableFramePool(ctx, (long) ctx.getInitialFrameSize() * (long) 2048);
         bufferManager = new FramePoolBackedFrameBufferManager(framePool);
         nsTable = new SerializableHashTable(NUM_PART, ctx, bufferManager);
     }
